Updated rigid body code and added imgui controls for rigid body calculations.

+++ Physics Framework/Physics/Collisions/ContactResolver.h
@@ -8,31 +8,26 @@ class ContactResolver
 {
 	friend class Contact;
 public:
-	ContactResolver( unsigned iterations, float velocityEps = ( float )0.01, float positionEps = ( float )0.01 );
-	ContactResolver( unsigned velocityIt, unsigned positionIt, float positionEp = ( float )0.01, float velocityEp = ( float )0.01 );
+	ContactResolver( uint32_t iterations, float velocityEps = 0.01f, float positionEps = 0.01f );
+	ContactResolver( uint32_t velocityIt, uint32_t positionIt, float positionEp = 0.01f, float velocityEp = 0.01f );
 
 	bool IsValid();
-	void SetIterations( unsigned velocityIt, unsigned positionIt );
-	void SetIterations( unsigned iterations );
+	void SetIterations( uint32_t velocityIt, uint32_t positionIt );
+	void SetIterations( uint32_t iterations );
 	void SetEpsilon( float velocityEp, float positionEp );
-	void ResolveContacts( Contact* contacts, unsigned numContacts, float dt );
-
+	void ResolveContacts( Contact* contacts, uint32_t numContacts, float dt );
 protected:
-	void AdjustVelocities( Contact* contacts, unsigned numContacts, float dt );
-	void AdjustPositions( Contact* contacts, unsigned numContacts, float dt );
-	void PrepareContacts( Contact* contacts, unsigned numContacts, float dt );
-
+	void AdjustVelocities( Contact* contacts, uint32_t numContacts, float dt );
+	void AdjustPositions( Contact* contacts, uint32_t numContacts, float dt );
+	void PrepareContacts( Contact* contacts, uint32_t numContacts, float dt );
 public:
-	unsigned velocityIterationsUsed;
-	unsigned positionIterationsUsed;
-
+	uint32_t velocityIterationsUsed;
+	uint32_t positionIterationsUsed;
 protected:
-	unsigned velocityIterations;
-	unsigned positionIterations;
-
+	uint32_t velocityIterations;
+	uint32_t positionIterations;
 	float velocityEpsilon = 0.01;
 	float positionEpsilon = 0.01;
-
 private:
 	bool validSettings;
 };
